+++ docs/CHANGELOG.md
@@ -10,6 +10,7 @@ Changelog
 -----------------------
 - Enh #48: Use RichText for file description
 - Enh #103: Allow to edit topics from the file edit form
+- Enh #82: Move files and folders by drag & drop
 
 
 0.13.1 - July 29, 2021